refactor(gateway): extract chat send setup phases
